use location.can_reach instead of state.locations_checked
I was running off of the assumption that locations_checked was populated 
when calculating access rules, when it really doesn't. After a night's 
sleep and looking at the debug logs for the generation, I saw that every 
progression item was being placed in the Easy tier of levels, because it 
could never reach further than that.

Sorry to the #manual-support channel for receiving my rambling while I 
was working this all out!